Elmer v0.0.12 Elmer

Elmer is a tool for generating common elm-lang structures. This module is for abstracting out common functions that many of the modules use.

Functions

ecto_defaults()

Returns a map of default ecto values

ecto_elm()

Returns a map of ecto types to elm types

prompt_path()