Elmer v0.0.12 Mix.Tasks.Elmer.Gen.Cmd

Mix task to generate Cmds.elm files.

Functions

run(args)

Creates a new elm Cmd in the directory corresponding to your .

Run with mix elmer.gen.cmd <module name> <model> <args>

Currently this module is only setup to handle REST Cmds but in the future it will be able to handle multiple types

Args are in the format “CmdName:Type:RequestType:FuncParams”

Example: mix elmer.gen.cmd Players Player CreatePlayer:rest:POST:Player

Type is “rest”, for now

RequestType is: “GET”, “PATCH”, “POST”, “DELETE”, etc.

FuncParams are what get plugged into your Cmd function