Elmer v0.0.6 Mix.Tasks.Elmer.Gen.Msg

This is a mix task for creating a Msg template.

Functions

run(args)

Creates a new elm Msg in the directory corresponding to your .

Run with mix elmer.gen.msg <module name> <args>

module name is the module you want to place this Msg in, for example: Players.Striker will create the file /Players/Striker/Msgs.elm

Args are similar to phoenix gen arguments where they are in the format “MsgName:Type1:Type2:TypeN”

MsgName is the name of the message you want to generate

The types will be entered in exactly as you type them.

You can enter in as many messages as you want as long as they’re separated by a space